requirements.txt
pytest==4.5.0
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest==2.8.6